Feat: Gmail 사용한 로그인 기능 구현 (#6)
* Feat: imap을 사용한 메일 불러오기 기능

* Feat: gmail 인증 테스트 코드 작성

* Feat: 회원가입 API 기본 구현

* Feat: verification code 요청 api

* Feat: 테스트 코드 작성중

* Feat: 로그인 기능 구현 (테스트 코드 X)

* Fix: 테스트 코드 수정

* Refactor: 코드 리팩토링 및 에러코드 변경

* Test: 로그인 테스트 코드

* Issue: 회원가입 API 테스트 오류 발생 해결중

* Fix: 회원가입 API 테스트 코드 수정 완료

* Feat: SignUpReq 수정

* Refactor: gmail 인증 관련 코드 리팩토링

* Fix: 로그인에서 Gmail 검증 누락 수정 및 테스트 코드 수정

+++ src/main/java/Dino/Duett/domain/signup/dto/SignUpReq.java
@@ -23,7 +23,7 @@ public class SignUpReq {
     @NotBlank
     private String birth;
     @NotNull
-    private int[] location;
+    private double[] location;
     @NotNull
     private MultipartFile profileImage;
     @NotBlank
